# Conflicts:
#	src/main/java/com/fitbum/dto/RoleDTO.java
#	src/main/java/com/fitbum/repositorios/MenuRepositorio.java
#	src/main/java/com/fitbum/repositorios/usuarios/UsuarioRepositorio.java
#	src/main/java/com/fitbum/servicios/usuarios/UsuarioServicio.java

+++ src/main/java/com/fitbum/repositorios/usuarios/RoleRepositorio.java
@@ -5,7 +5,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepositorio extends JpaRepository<Role, Integer>{
-Page<Role> findAll(Pageable pageable);
+
+public interface RoleRepositorio extends JpaRepository<Role, Long>{
+
+    Page<Role> findAll(Pageable pageable);
 
 }
